requirements.txt
numpy==2.2.6
torch==2.9.0
torchvision==0.24.0
opencv-python==4.12.0.88
Pillow==12.0.0
scikit-learn==1.7.2
imutils==0.5.4
matplotlib==3.10.7
yagmail==0.15.293
streamlit